requirements.txt
streamlit
matplotlib
numpy
Pillow
opencv-python
easyocr